qa & prod branches have been merged

+++ src/app/util/landing-pages-list-and-grid-view/landing-pages-list-and-grid-view.component.ts
@@ -23,6 +23,8 @@ import { CopyDto } from '../models/copy-dto';
 import { Properties } from 'app/common/models/properties';
 import { LandingPageShareDto } from 'app/dashboard/user-profile/models/LandingPageShareDto';
 import { PartnerCompanyAndGroupsComponent } from '../partner-company-and-groups/partner-company-and-groups.component';
+import { XAMPLIFY_CONSTANTS } from 'app/constants/xamplify-default.constants';
+import { Location } from '@angular/common';
 declare var swal: any, $: any;
 @Component({
   selector: 'app-landing-pages-list-and-grid-view',
@@ -77,12 +79,13 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
   @Output() viewVendorPageAnalytics = new EventEmitter();
   @Input() isMasterLandingPages =  false;
   @Output() isFormAnalytics = new EventEmitter();
+  @Input() welcomePages = false;
   @ViewChild("partnerCompanyAndGroupsComponent") partnerCompanyAndGroupsComponent:PartnerCompanyAndGroupsComponent;
   constructor(public referenceService: ReferenceService,public httpRequestLoader: HttpRequestLoader, public pagerService:PagerService, public authenticationService: AuthenticationService,
       public router: Router, public landingPageService: LandingPageService, public logger: XtremandLogger,
       public actionsDescription: ActionsDescription, public sortOption: SortOption,
       private utilService: UtilService, private route: ActivatedRoute,public renderer:Renderer,
-      private vanityUrlService:VanityURLService,public properties:Properties, private changeDetectorRef: ChangeDetectorRef,) {
+      private vanityUrlService:VanityURLService,public properties:Properties, private changeDetectorRef: ChangeDetectorRef,public location: Location) {
         this.pagination.vanityUrlFilter =this.vanityUrlService.isVanityURLEnabled();
         this.loggedInUserId = this.authenticationService.getUserId();
         this.referenceService.renderer = this.renderer;
@@ -110,7 +113,7 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
 }
 
   private sefDefaultViewType() {
-    if(this.vendorJourney || this.isLandingPages || this.isMasterLandingPages){
+    if(this.vendorJourney || this.isLandingPages || this.isMasterLandingPages || this.welcomePages){
         this.viewType = 'g';
     }else if (this.folderListViewCategoryId != undefined) {
         this.categoryId = this.folderListViewCategoryId;
@@ -176,6 +179,13 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
         this.pagination.loginAsUserId = this.loggedInUserId;
         this.pagination.companyId = this.loggedInUserCompanyId;
         this.pagination.masterLandingPage = true;
+      }else if(this.welcomePages){
+        this.pagination.source = "WELCOME_PAGE";
+        this.pagination.defaultLandingPage = false;
+        this.pagination.loginAsUserId = this.loggedInUserId;
+        this.pagination.companyId = this.loggedInUserCompanyId;
+        this.pagination.welcomePages = true;
+
       }else{
         this.pagination.source = "MANUAL";
       }
@@ -297,7 +307,11 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
               confirmButtonText: 'Yes, delete it!'
 
           }).then(function () {
-              self.deleteById(landingPage);
+            if(self.welcomePages){
+                self.welcomePageDelete(landingPage);
+            }else{
+                self.deleteById(landingPage);
+            }
           }, function (dismiss: any) {
               console.log('you clicked on option' + dismiss);
           });
@@ -309,7 +323,7 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
 
 
   editLandingPage(id: number) {
-    if(this.vendorJourney || this.isLandingPages || this.isMasterLandingPages){
+    if(this.vendorJourney || this.isLandingPages || this.isMasterLandingPages || this.welcomePages){
         this.landingPageService.getById(id).subscribe(
             (data: any) => {
                 this.vendorLandingPage.emit(data.data);
@@ -392,7 +406,7 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
   }
 
   goToFormAnalytics(id: number) {
-    if(this.isMasterLandingPages || this.vendorJourney){
+    if(this.isMasterLandingPages || this.vendorJourney || this.welcomePages){
         this.isFormAnalytics.emit(id);
     }else{
       if(this.categoryId>0){
@@ -407,7 +421,7 @@ export class LandingPagesListAndGridViewComponent implements OnInit,OnDestroy {
       this.router.navigate(['/home/forms/partner/lf/' + alias]);
   }
   goToLandingPageAnalytics(id: number) {
-    if(this.vendorJourney || this.isMasterLandingPages){
+    if(this.vendorJourney || this.isMasterLandingPages || this.welcomePages){
         this.viewAnalytics.emit(id);
     }else{
       if(this.categoryId>0){
@@ -468,7 +482,7 @@ copy(landingPage:any){
 
   findExistingPageNames(landingPage:any){
     this.ngxloading = true;
-    this.landingPageService.getAvailableNames(this.loggedInUserId).subscribe(
+    this.landingPageService.getAvailableNames(this.loggedInUserId, this.welcomePages).subscribe(
       (data: any) => {
           let pageNames = data;
           this.copyModalPopupComponent.openModalPopup(landingPage.id,landingPage.name,"Page",pageNames);
@@ -586,4 +600,104 @@ copy(landingPage:any){
     publishVendorJourney(){
         this.partnerCompanyAndGroupsComponent.publish();
     }
+
+    publishWelcomePage(landingPageId:number){
+        this.customResponse = new CustomResponse();
+        this.selectedLandingPageId = landingPageId;
+        let landingPage = new LandingPage();
+        landingPage.id = landingPageId;
+            this.ngxloading = true;
+            this.landingPageService.updateWelcomePage(landingPage).subscribe(
+              (response) => {
+                this.customResponse = new CustomResponse('SUCCESS', response.message, true);
+                  this.ngxloading = false;
+              },
+              error => {
+                this.ngxloading = false;
+                this.logger.errorPage(error);
+              }, ()=>{
+                this.updateWelcomePageActiveKeyValueOrFinsLandingPages(true);
+              });
+    }
+
+
+
+    unPublishWelcomePage(landingPageId:number){
+        this.customResponse = new CustomResponse();
+        this.selectedLandingPageId = landingPageId;
+        let landingPage = new LandingPage();
+        landingPage.id = landingPageId;
+            this.ngxloading = true;
+            let self = this;
+            this.landingPageService.unPublishWelcomePage(landingPage).subscribe(
+              (response) => {
+                this.customResponse = new CustomResponse('SUCCESS', response.message, true);
+                self.ngxloading = false;
+              },
+              error => {
+                this.ngxloading = false;
+                this.logger.errorPage(error);
+              }, ()=>{
+                this.updateWelcomePageActiveKeyValueOrFinsLandingPages(false);
+
+              });
+    }
+
+    private updateWelcomePageActiveKeyValueOrFinsLandingPages(isWelcomePageActivated:boolean) {
+        if (this.authenticationService.vanityURLEnabled) {
+            this.updateWelcomePageActiveKeyAndReload(isWelcomePageActivated);
+        } else {
+            this.findLandingPagesWithPageIndexOne();
+        }
+    }
+    updateWelcomePageActiveKeyAndReload(isWelcomePageActivated:boolean){
+        let currentUser = this.authenticationService.getLocalStorageItemByKey(XAMPLIFY_CONSTANTS.currentUser);
+        currentUser[XAMPLIFY_CONSTANTS.welcomePageEnabledKey] = isWelcomePageActivated;
+        //try to set only the welcomePageEnabledKey value insted of the whole 'currentUser'
+        localStorage.setItem('currentUser',JSON.stringify(currentUser)) ;
+
+        this.ngxloading = true;
+        if(currentUser[XAMPLIFY_CONSTANTS.welcomePageEnabledKey]){
+            this.location.replaceState('/welcome-page');
+        }else{
+            this.location.replaceState('/home/dashboard');
+        }  
+        window.location.reload();       
+        this.ngxloading = false;
+    }
+
+    welcomePageDelete(landingPage: LandingPage) {
+        this.customResponse = new CustomResponse();
+        this.referenceService.loading(this.httpRequestLoader, true);
+        this.referenceService.goToTop();
+        this.landingPageService.welcomePageDeletebById(landingPage.id)
+            .subscribe(
+                (response: any) => {
+                    if(response.access){
+                        if (response.statusCode == 200) {
+                            let message = landingPage.name + " deleted successfully";
+                            this.customResponse = new CustomResponse('SUCCESS', message, true);
+                            this.findLandingPagesWithPageIndexOne();
+                        } else {
+                            let pageNames = "";
+                            $.each(response.data, function (index:number, value:any) {
+                                pageNames += (index + 1) + ". " + value + "\n\n";
+                            });
+                            let message = response.message + "\n\n" + pageNames;
+                            this.customResponse = new CustomResponse('ERROR', message, true);
+                            this.referenceService.loading(this.httpRequestLoader, false);
+                        }
+                    }else{
+                        this.authenticationService.forceToLogout();
+                    }
+                   
+  
+                },
+                (error: string) => {
+                    this.referenceService.showServerErrorMessage(this.httpRequestLoader);
+                    this.customResponse = new CustomResponse('ERROR', this.httpRequestLoader.message, true);
+                }
+            );
+    }
 }
+
